requirements.txt
# 🚀 AGENTE IA IMOBILIÁRIO - Dependências
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Requisições HTTP
requests==2.31.0

# Variáveis de ambiente
python-dotenv==1.0.0

# Supabase (já temos)
supabase==2.3.4

# OpenAI (para futuras implementações)
openai==1.3.7

# Google Cloud Vision (para futuras implementações)
google-cloud-vision==3.4.5

# Utilitários
python-multipart==0.0.6

# Para desenvolvimento e testes
pytest==7.4.3
pytest-asyncio==0.21.1 